hay_v2_bot/requirements.txt
openai
python-dotenv
pinecone
pyTelegramBotAPI
haystack-ai
pinecone-haystack
docling
transformers
protobuf
requests
docstring-parser
jsonschema
